Convert TextField to tsx
