+++ packages/design-system-docs/src/pages/components/TextField/_Mask.docs.scss
@@ -17,7 +17,7 @@ value, you can import and call the `unmaskValue` method.
 
 @react-example Mask.example.jsx
 
-@react-props TextField.jsx
+@react-props TextField.tsx
 
 Style guide: components.masked-field.react
 */
